requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.6.4
python-dotenv==1.0.0
jinja2
python-multipart
aiofiles
google-cloud-firestore
